solving codes issue
