+++ src/strategic_classification/models/vanila/__init__.py
@@ -1,0 +1,1 @@
+from .vanila import train_vanila
